+++ src/page-objects/index.ts
@@ -1,6 +1,8 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 export { default as BasePageObject } from './base';
-export { default as ScreenshotPageObject, PermutationScreenshot, RawPermutationScreenshot } from './screenshot';
+export { default as ScreenshotBasePageObject } from './screenshot-base';
+export { default as ScreenshotPageObject, PermutationScreenshot } from './screenshot';
+export { default as RawScreenshotPageObject, RawPermutationScreenshot } from './raw-screenshot';
 export { default as EventsSpy } from './events-spy';
-export { Screenshot, RawScreenshot, ScreenshotWithOffset, ElementSize, ElementRect, ElementOffset } from './types';
+export { ScreenshotWithOffset, RawScreenshot, ElementSize, ElementRect, ElementOffset } from './types';
