+++ src/page-objects/types.ts
@@ -26,29 +26,21 @@ export interface ViewportSize extends ElementOffset, ElementSize {
 
 export interface ScreenshotCapturingOptions {
   viewportOnly?: boolean;
-  singleElements?: boolean;
 }
 
-/**
- * A raw screenshot with base64 data and dimensions. No decoded image, no offset.
- * Returned when singleElements is true (takeElementScreenshot was used).
- */
-export interface RawScreenshot extends ElementSize {
-  image?: PNG;
-  rawBase64: string;
-  pixelRatio?: number;
-}
-
-/**
- * A decoded screenshot with image data and offset for cropping.
- * Returned when singleElements is false/absent (full-page screenshot path).
- * This is the original type consumers expect.
- */
 export interface ScreenshotWithOffset extends ElementSize {
   image: PNG;
   offset: ElementOffset;
   pixelRatio?: number;
+  /** Optional raw base64 PNG for fast byte-equality comparison. */
+  rawBase64?: string;
 }
 
-/** Union of both screenshot types. */
-export type Screenshot = RawScreenshot | ScreenshotWithOffset;
+/**
+ * A raw screenshot with base64 data and dimensions. No decoded image, no offset.
+ * Returned by RawScreenshotPageObject which uses takeElementScreenshot.
+ */
+export interface RawScreenshot extends ElementSize {
+  rawBase64: string;
+  pixelRatio?: number;
+}
