+++ test/page-object.test.ts
@@ -49,15 +49,6 @@ test(
 );
 
 test(
-  'focusNextElement',
-  setupTest(async page => {
-    await page.click('#input-1');
-    await page.focusNextElement();
-    expect(await page.isFocused('#input-2')).toBe(true);
-  })
-);
-
-test(
   'isSelected',
   setupTest(async page => {
     expect(await page.isSelected('#checkbox')).toBe(true);
@@ -270,31 +261,6 @@ test(
 );
 
 test(
-  'scrollToRight',
-  setupTest(async page => {
-    const width = 400;
-    const overscroll = (width * 20) / 100; // The container has 120% width
-    await page.setWindowSize({ width, height: 300 });
-    await page.scrollToRight('#scrollable-container');
-    expect(await page.getElementScroll('#scrollable-container')).toEqual({ top: 0, left: overscroll });
-  })
-);
-
-test(
-  'scrollToBottom',
-  setupTest(async page => {
-    const testElement = '#scrollable-container';
-    const { height } = await page.getBoundingBox(testElement);
-    const scrollHeight = (await page.getElementProperty(testElement, 'scrollHeight')) as number;
-
-    await page.scrollToBottom(testElement);
-
-    const { top } = await page.getElementScroll(testElement);
-    expect(top).toBeGreaterThanOrEqual(scrollHeight - height);
-  })
-);
-
-test(
   'getBoundingBox',
   setupTest(async page => {
     const box = await page.getBoundingBox('#text-content');
@@ -362,83 +328,3 @@ test(
     });
   })
 );
-
-describe('capturePermutations', () => {
-  test(
-    'captures all permutations with correct ids and dimensions',
-    setupTest(async page => {
-      const permutations = await page.capturePermutations();
-
-      expect(permutations).toHaveLength(3);
-      expect(permutations.map(p => p.id)).toEqual(['perm-1', 'perm-2', 'perm-3']);
-
-      permutations.forEach(perm => {
-        expect(perm.image).toBeDefined();
-        expect(perm.width).toBeGreaterThan(0);
-        expect(perm.height).toBeGreaterThan(0);
-        expect(perm.offset.top).toBeGreaterThanOrEqual(0);
-        expect(perm.offset.left).toBeGreaterThanOrEqual(0);
-      });
-
-      // Verify permutations are in order (top to bottom)
-      expect(permutations[0].offset.top).toBeLessThan(permutations[1].offset.top);
-      expect(permutations[1].offset.top).toBeLessThan(permutations[2].offset.top);
-    }, './test-permutations.html')
-  );
-
-  test(
-    'throws error when no permutations found',
-    setupTest(async page => {
-      await expect(page.capturePermutations()).rejects.toThrowError('No permutations found on current page.');
-    })
-  );
-});
-
-describe('singleElements option', () => {
-  test(
-    'captureBySelector with singleElements returns rawBase64 without image',
-    setupTest(async page => {
-      const result = await page.captureBySelector('#text-content', { singleElements: true });
-
-      expect(result.rawBase64).toBeDefined();
-      expect(result.rawBase64.length).toBeGreaterThan(0);
-      expect(result.width).toBeGreaterThan(0);
-      expect(result.height).toBeGreaterThan(0);
-      expect('image' in result).toBe(false);
-      expect('offset' in result).toBe(false);
-    })
-  );
-
-  test(
-    'captureViewport with singleElements returns rawBase64 without image',
-    setupTest(async page => {
-      const result = await page.captureViewport({ singleElements: true });
-
-      expect(result.rawBase64).toBeDefined();
-      expect(result.rawBase64.length).toBeGreaterThan(0);
-      expect(result.width).toBeGreaterThan(0);
-      expect(result.height).toBeGreaterThan(0);
-      expect('image' in result).toBe(false);
-      expect('offset' in result).toBe(false);
-    })
-  );
-
-  test(
-    'capturePermutations with singleElements returns rawBase64 without image',
-    setupTest(async page => {
-      const permutations = await page.capturePermutations({ singleElements: true });
-
-      expect(permutations).toHaveLength(3);
-      expect(permutations.map(p => p.id)).toEqual(['perm-1', 'perm-2', 'perm-3']);
-
-      permutations.forEach(perm => {
-        expect(perm.rawBase64).toBeDefined();
-        expect(perm.rawBase64.length).toBeGreaterThan(0);
-        expect(perm.width).toBeGreaterThan(0);
-        expect(perm.height).toBeGreaterThan(0);
-        expect('image' in perm).toBe(false);
-        expect('offset' in perm).toBe(false);
-      });
-    }, './test-permutations.html')
-  );
-});
